+++ tests/unit/test_markdown.py
@@ -119,6 +119,18 @@ class TestMarkdownFormat:
         )
 
     @staticmethod
+    def should_generate_possible_values_and_examples(runner: CliRunner, mocker: MockerFixture):
+        expected_string = (
+            "# `possible_values_and_examples`\n\n"
+            "**required**\n\n"
+            "## possible values\n\n"
+            "- `debug`: debug level\n- `info`: info level\n\n"
+            "## examples\n\n"
+            "- `debug`: debug level\n- `info`: info level\n\n"
+        )
+        assert expected_string in run_app_with_settings(mocker, runner, ExamplesSettings)
+
+    @staticmethod
     @pytest.mark.parametrize(
         "settings_class",
         [